DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors
(cherry picked from commit 6f9731591ea31a6fc6fd48d29b87c5eff53e2a98)

DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors

(cherry picked from commit 35812e50230f2c5e408ed1bb1040c038af4d21eb)

DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors

(cherry picked from commit 53c51351005909c55dbe67af7687e582b28ad769)

DELIA-57501 - Fix Warnings in RDKServices builds and enable warnings as errors

Moving these flags from here to rdkservices_git.bb

(cherry picked from commit 273c8183c8c0fbe0df54b06cbd2954819192f18c)
(cherry picked from commit 25d0e0f6324eb55246a73f4f41c3f95853ab0ab9)

DELIA-57501 - Fix warnings

(cherry picked from commit 3b8ef66dbca5c055dcc5f5240b821f1bbb9a9440)

DELIA-57501 - Fix Warnings in RDKServices

Reverted whitespace changes

(cherry picked from commit 592379d411a74e9332d268a66de7966f4b31a2e9)

DELIA-57501-Changed typecasts

(cherry picked from commit 40ee995e6587a0440b558f23def02f1e5f0049ed)

DELIA-57501 - some more warning fixes (warehouse)

(cherry picked from commit fe78c15a11e6703f8845ae10b53659a517286523)

DELIA-57501 - Fix warnings

Removed plugin specific warning flags. These will be fixed later.

(cherry picked from commit 5df75427cbd8f470877c1d82e6ca46bfaf2fd00c)

DELIA-57833 - Fix Warnings

Fix Warnings, suppress if the warnings coming from outside of rdkservices.

(cherry picked from commit 4f32d910617bdbedf9b64d29cf9e87ae4d60a35c)

DELIA-57833 - Fix Warnings - suppress macro redefine errors

(cherry picked from commit 6c424853404a791d3e23b81e56a1af3b5ce2ac6c)

DELIA-57833 - Fix Warnings - LOGWARN

(cherry picked from commit 4225ca8ba9ff8c7c1564f43d77ab86041bde968c)

DELIA-57833 - Fix Warnings in RDKServices and enable warnings as errors - 2

(cherry picked from commit 1972ada597be99fcfa6df9d5a4ea461eda8b4207)

DELIA-57833 - Fix Warnings in RDKServices and enable warnings as errors - 2

(cherry picked from commit 1ef9d5f17fb899e9d875cc8297c5871033c2f09c)

DELIA-57833 - Fix Warnings in RDKServices and enable warnings as error

(cherry picked from commit ef001b96844aba09cbaa89d67ac9fe7fa5241f52)

DELIA-57833 - Fix Warnings in RDKServices and enable warnings as error

(cherry picked from commit f29276ec5bf02b4ea9b65b4005831d7cad65f9fd)

+++ WifiManager/impl/WifiManagerSignalThreshold.cpp
@@ -119,7 +119,7 @@ void WifiManagerSignalThreshold::setSignalThresholdChangeEnabled(bool enabled, i
     {
         int64_t number = std::stoi(response["state"].String());
         state = (WifiState) number;
-        LOGINFO("wifi state = %d", state);
+        LOGINFO("wifi state = %d", static_cast<int>(state));
     }
     else
     {
